Fix visual studio errors and warnings

+++ src/simulation/elements/COAL.cpp
@@ -71,7 +71,7 @@ int COAL_update(UPDATE_FUNC_ARGS)
 	return 0;
 }
 
-#define FREQUENCY 3.1415/(2*300.0f-(300.0f-200.0f))
+#define FREQUENCY 3.1415f/(2*300.0f-(300.0f-200.0f))
 int COAL_graphics(GRAPHICS_FUNC_ARGS)
 {
 	*colr += (int)((cpart->tmp2-295.15f)/3);
@@ -85,11 +85,11 @@ int COAL_graphics(GRAPHICS_FUNC_ARGS)
 
 	if (cpart->temp > 395.15f)
 	{
-		int q = (cpart->temp > 595.15f) ? 200.0f : cpart->temp - 395.15f;
+		int q = (cpart->temp > 595.15f) ? 200 : (int)(cpart->temp - 395.15f);
 
-		*colr += sin(FREQUENCY*q) * 226;
-		*colg += sin(FREQUENCY*q*4.55 + 3.14) * 34;
-		*colb += sin(FREQUENCY*q*2.22 + 3.14) * 64;
+		*colr += (int)(sin(FREQUENCY*q) * 226);
+		*colg += (int)(sin(FREQUENCY*q*4.55 + 3.14) * 34);
+		*colb += (int)(sin(FREQUENCY*q*2.22 + 3.14) * 64);
 	}
 	return 0;
 }
